Move month mapping to utilities.
