+++ src/soep_preparation/clean_variables/biobirth.py
@@ -2,7 +2,7 @@
 
 import pandas as pd
 
-from soep_preparation.clean_variables import month_mapping
+from soep_preparation.utilities import month_mapping
 from soep_preparation.utilities.data_manipulator import (
     float_to_int,
     object_to_int,
